tests/wpt/tests/html/canvas/offscreen/path-objects/2d.path.arc.angle.1.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<meta charset="UTF-8">
<title>OffscreenCanvas test: 2d.path.arc.angle.1</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>

<h1>2d.path.arc.angle.1</h1>
<p class="desc">arc() draws pi/2 .. -pi anticlockwise correctly</p>


<script>
var t = async_test("arc() draws pi/2 .. -pi anticlockwise correctly");
var t_pass = t.done.bind(t);
var t_fail = t.step_func(function(reason) {
    throw reason;
});
t.step(function() {

  var canvas = new OffscreenCanvas(100, 50);
  var ctx = canvas.getContext('2d');

  ctx.fillStyle = '#0f0';
  ctx.fillRect(0, 0, 100, 50);
  ctx.fillStyle = '#f00';
  ctx.beginPath();
  ctx.moveTo(100, 0);
  ctx.arc(100, 0, 150, Math.PI/2, -Math.PI, true);
  ctx.fill();
  _assertPixel(canvas, 50,25, 0,255,0,255);
  t.done();

});
</script>
